+++ client/views/Facility/FacilityWidget.jsx
@@ -1,0 +1,12 @@
+
+FacilityWidget= React.createClass({
+	render() {
+		return (
+			<FlipWidget
+				front={FacilityEdit}
+				back={FacilityView}
+				item={this.props.item}
+			/>
+		)
+	}
+});