+++ client/views/Issue/IssueDetail.jsx
@@ -122,6 +122,7 @@ IssueDynamicArea = React.createClass({
         return (
             <div className="row">
                 <div className="col-lg-12">
+                    <a href={FlowRouter.path('request',{_id:issue._id})}>Link</a>
                     <div style={{borderTop:"1px solid #ccc",marginTop:"10px",paddingBottom:"10px"}}></div>
                     <span className="btn btn-sm btn-flat issue-nav-btn">Description</span><br/>
                     <div className="issue-dynamic-area">
@@ -145,8 +146,7 @@ IssueDynamicArea = React.createClass({
                         },{
                             tab:<span><span>Updates</span>{this.data.messageCount?<span className="label label-notification">{this.data.messageCount}</span>:null}</span>,
                             content:<div>
-                                {/*<IssueDiscussion items={notifications}/>*/}
-                                <NewsFeed feedId={issue.getFeedId()} feedName={issue.getFeedName()}/>
+                                <Inbox for={issue} />
                             </div>
                         }
                     ]} />