+++ client/views/Message/MessageView.jsx
@@ -1,18 +1,19 @@
 
-NewsPost = React.createClass({
+MessageView = React.createClass({
 
     mixins: [ReactMeteorData],
 
     getMeteorData() {
-        var query, post, creator;
+        var query, message, creator;
         query = this.props.item;
-        post = Posts.findOne(query);
-        if(post) {
-            creator = post.getCreator()
+        message = Posts.findOne(query);
+        if(message) {
+            creator = message.getCreator()
         }
         return {
             creator:creator,
-            post:post
+            inbox:this.props.inbox,
+            message:message
         }
     },
 
@@ -22,9 +23,10 @@ NewsPost = React.createClass({
 
     submit() {
         var input = this.refs.input;
+        var inboxId = this.data.inbox.getInboxId();
         Meteor.call("Posts.new",{
-            feedId:this.props.feedId,
-            subject:"posted to "+this.props.feedName,
+            inboxId:inboxId,
+            verb:"sent a message to",
             body:input.value
         });
         input.value = null;
@@ -38,20 +40,25 @@ NewsPost = React.createClass({
     },
 
     render() {
-        var post = this.data.post||{};
+        var message = this.data.message||{};
         var creator = this.data.creator||Meteor.user();
-        var createdAt = post.createdAt;
+        var createdAt = message.createdAt;
         var used = false;
         return(
             <div>
                 <ContactAvatarSmall item={creator}/>
                 <div className="media-body">
-                    {post.body||post.subject?
+                    {message.body||message.subject||message.verb?
                     <div>
-                        <small className="pull-right">{moment(post.createdAt).fromNow()}</small>
-                        <strong>{creator.getName()}</strong> {post.subject}<br/>
+                        <small className="pull-right">{moment(message.createdAt).fromNow()}</small>
+                        <strong>{creator.getName()}</strong> {
+                        message.verb?
+                            <span>{message.verb} <b><a href={message.getTargetUrl()}>{message.getTargetName()}</a></b></span>
+                        :
+                            <span>{message.subject}</span>
+                        }<br/>
                         <small className="text-muted">{moment(createdAt).format('MMM Do YYYY, h:mm:ss a')}</small>
-                        <div>{post.body}</div>
+                        <div>{message.body}</div>
                     </div>
                     :
                     <textarea 
@@ -59,7 +66,7 @@ NewsPost = React.createClass({
                         style={{width:"80%"}}
                         placeholder="Leave a comment or question..."
                         className={"input inline-form-control "+(used?'used':'')}
-                        defaultValue={post.body} 
+                        defaultValue={message.body} 
                         onKeyDown={this.handleKeyPress}>
                     </textarea>
                     }