+++ client/views/Message/NotificationView.jsx
@@ -3,28 +3,33 @@ NotificationSummary = React.createClass({
     mixins: [ReactMeteorData],
 
     getMeteorData() {
-        var query, post, creator;
+        var query, message, creator;
         query = this.props.item;
-        post = Posts.findOne(query);
-        if(post) {
-            creator = post.getCreator()
+        message = Posts.findOne(query);
+        if(message) {
+            creator = message.getCreator()
         }
         return {
             creator:creator,
-            post:post
+            message:message
         }
     },
 
 	render() {
-        var post = this.data.post||{};
+        var message = this.data.message||{};
         var creator = this.data.creator||Meteor.user();
-        var createdAt = post.createdAt;
+        var createdAt = message.createdAt;
 		return (
             <div>
             	<ContactAvatarSmall item={creator}/>
-                <small>{moment(post.createdAt).fromNow()}</small>
+                <small>{moment(message.createdAt).fromNow()}</small>
                 <div>
-                    <strong>{creator.getName()}</strong> {post.subject}<br/>
+                    <strong>{creator.getName()}</strong> {
+                    message.verb?
+                        <span>{message.verb} <b><a href={message.getTargetUrl()}>{message.getTargetName()}</a></b></span>
+                    :
+                        <span>{message.subject}</span>
+                    }<br/>
                 </div>
             </div>
 		)