+++ lib/_namespace.jsx
@@ -73,45 +73,25 @@ FM.createCollection = function(name,template,shouldNotCreateSchema) {
 		if(!item.createdAt) {
 			item.createdAt = moment().toDate();
 		}
-		/*
-		console.log({
-			saving:item
-		});
-		*/
 		var response = collection.upsert(item._id, {$set: _.omit(item, '_id')});
 		return {
 			_id:response.insertedId
 		}
 	}
 	methods[name+'.destroy'] = function(item) {
-		/*console.log({
-			destroying:item
-		});*/
 		collection.remove(item._id);
 	}
-	methods[name+'.new'] = function(item,actor) {
-		actor = actor||Meteor.user();
+	methods[name+'.new'] = function(item) {
+		var owner = Meteor.user();
 		newItem = _.extend({
 			isNewItem:true,
 			creator:{
-				_id:actor._id,
-				name:actor.name,
-				thumb:actor.profile.thumb
+				_id:owner._id,
+				name:owner.name,
+				thumb:owner.profile.thumb
 			},
 		},newItemTemplate(item),item);
-		/*
-		console.log({
-			creating:newItem,
-			with:item,
-			and:newItemTemplate
-		});*/
-
-		return collection.insert(newItem,function(err,obj){
-			/*console.log({
-				error:err,
-				data:obj
-			});*/			
-		});
+		return collection.insert(newItem);
 	}
 	methods[name+'.getTemplate'] = function(item) {
 		return newItemTemplate(item);
@@ -122,7 +102,7 @@ FM.createCollection = function(name,template,shouldNotCreateSchema) {
 	collection.helpers({
 		collectionName:name,
 		defaultThumbUrl:"img/default-placeholder.png",
-		save(extension) {
+		save(extension,callback) {
 			console.log('calling save method...');
 			var obj = this;
 			if(extension) {
@@ -130,6 +110,7 @@ FM.createCollection = function(name,template,shouldNotCreateSchema) {
 					obj[i] = extension;
 				}
 			}
+			Meteor.call(name+'.save',obj,callback);
 		},
 		isNew() {
 			// now this is a case where we should have underscore prefix
@@ -151,12 +132,34 @@ FM.createCollection = function(name,template,shouldNotCreateSchema) {
 		getName(){
 			return this.name;
 		},
+		getInboxName() {
+    		return this.getName()+"'s"+" inbox";
+  		},
+		getInboxId() {
+			return {
+				collectionName:this.collectionName,
+				name:this.getInboxName(),
+				path:this.path,
+				query:{_id:this._id}
+			}
+		},
+		getMessages() {
+		    return Posts.find({inboxId:this.getInboxId()}).fetch();
+  		},
+		getNotifications() {
+		    return Posts.find({inboxId:this.getInboxId()}).fetch();
+  		},
+		getMessageCount() {
+    		return Posts.find({inboxId:this.getInboxId()}).count();
+		},  		
+  		/*
 		getNotifications() {
 		    return Log.find({
 				"action.collectionName":name,
 				"action.query":{_id:this._id}
 		    }).fetch();
 		},
+		*/
 		getCreator() {
 			return Users.findOne(this.creator._id);
 		},